requirements.txt
pystac-client>=0.7
planetary-computer>=1.0
rasterio>=1.3
numpy>=1.24
matplotlib>=3.7
Pillow>=10.0
tqdm>=4.65
mercantile>=1.2
